k k milaya birse

+++ app/Http/Controllers/UserController.php
@@ -3,9 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreUserRequest;
+use App\Http\Requests\UpdateUserRequest;
 use App\Models\Role;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller {
     /**
@@ -14,6 +18,7 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function index() {
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $users = User::all();
         return view('users.index', compact('users'));
     }
@@ -24,6 +29,8 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function create() {
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $roles = Role::all();
         return view('users.create', compact('roles'));
     }
@@ -35,6 +42,7 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function store(StoreUserRequest $request) {
+
         // dd($request->all());
 
         $user = User::create($request->all());
@@ -49,6 +57,8 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function show(User $user) {
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         //
     }
 
@@ -59,8 +69,9 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function edit(User $user) {
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if($user->id == Auth::user()->id, Response::HTTP_FORBIDDEN, '403 Forbidden');
         $roles = Role::all();
-
         return view('users.edit', compact('user', 'roles'));
     }
 
@@ -71,8 +82,16 @@ class UserController extends Controller {
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, User $user) {
-        dd($request->all());
+    public function update(UpdateUserRequest $request, User $user) {
+        if ($request->password == null) {
+            $user->fill($request->except(['_token', 'role', 'password']));
+        } else {
+            $user->fill($request->except(['_token', 'role']));
+        }
+        $user->save();
+        $user->roles()->detach();
+        $user->roles()->attach($request->role);
+        return redirect(route('users.index'));
     }
 
     /**
@@ -82,6 +101,7 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function destroy(User $user) {
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $user->roles()->detach();
         $user->delete();
